pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexaroa"
dynamic = ["version"]
description = "Decentralized AI Network"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "NeuroShard Team", email = "team@neuroshard.com"}
]
keywords = [
    "ai", "machine-learning", "distributed-computing", "blockchain", 
    "llm", "neural-network", "decentralized", "training"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
    "safetensors>=0.3.0",
    "cryptography>=41.0.0",
    "mnemonic>=0.20",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
# GUI dependencies (optional - for desktop app)
gui = [
    "customtkinter>=5.2.0",
    "Pillow>=9.0.0",
    "pystray>=0.19.0",
]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]

[project.urls]
Homepage = "https://neuroshard.com"
Documentation = "https://docs.neuroshard.com"
Repository = "https://github.com/neuroshard/neuroshard"
Issues = "https://github.com/neuroshard/neuroshard/issues"

[project.scripts]
# Main entry point - run a node
neuroshard-node = "neuroshard.cli:main"
# Alias for convenience
neuroshard = "neuroshard.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "neuroshard.version.__version__"}

[tool.setuptools.package-data]
"neuroshard.ui" = ["templates/*.html", "static/**/*"]
"protos" = ["*.proto"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long - handled by black

